feat(ehr): elation patient router to appointment router
Ref: #1040
